rdma/ucma: Integrate rdma cm into ioctl framework
This provides an example of integrating the rdma cm into the
generic ioctl framework, along with porting the command used
to create an id.

A real solution would require porting all of the commands,
most of which would be straightforward.  The difficulting is
adopting the existing event interface to support events from
other name spaces.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>