rdma/ucm: Sketch for an ioctl framework
The following is version 2 of an outline for an ioctl framework.
This framework targets the rdma cm, but is intended to be expanded
to include verbs and other rdma devices, such as usnic, qib, and hfi1.
This series includes a patch that ports the rdma cm create id
command to use the ioctl framework.

The purpose of the patch is to drive discussion and feedback on
various design decisions, so that a single merged solution can be
developed.  The code is entirely untested and will not work without
completing the port of the existing write interface to ioctls.

The general architecture behind the implementation is described below.

Ioctls are logically grouped into name spaces.  Each name
space defines a set of ioctl handlers.  Ioctls are routed through the
framework, to a specific name space, and from there to handler routines
for that ioctl.  A name space may be associated with a specific kernel
driver, a hardware device, or a set of hardware resources on a device.
This series defines 2 name spaces:

- A name space manager, which is used to retrieve data about the
  name spaces that are available.
- The rdma ucm, which exports the rdma cm interfaces to user space.

(Note: I selected the term name space to correspond to programming
name spaces.  However, since Linux uses the same term for a different
purpose, the name should probably change.)


A name space is identified in the kernel by a structure with the
following fields:

uda_ns
======
id	Name space manager assigned identifier.  Note that the name
	space manager is always at id 0.
flags	Used to indicate when a name space has been closed and
	the underlying implementation (i.e. driver or device) is
	no longer available.
idr	Used to map kernel resources back to user space.
lock	Protects access to the idr
ioctl_base
	Name space have unique opcode values.  Each name space has a
	hard coded limit of 128 opcodes.  The ioctl_base is the value
	of the starting opcode.  A name space that requires more than
	128 opcodes will need to span 2 'name spaces'.
num_ioctls
	Total number of function handlers supported by the name space
[get_]ioctl_desc()
	Handler that takes as input an ioctl, verifies that the
	ioctl is properly formatted, including all arguments are
	valid, and returns a descriptor that describes how to
	process the ioctl.
close()	Generic close routine to release any underlying resources
	associated with an object.  This is invoked by the framework
	to cleanup resources when the name space is removed from the
	system
name	Name spaces are identified by a character string.


A primary goal of a name space is to associate an ioctl with a
descriptor.  The descriptor tells the framework how to process the
ioctl.  A descriptor is a simple structure with these fields:

flags	Specifies if the operation creates a new object, destroys
	an object, and the type of access required (shared or
	exclusive) of the object.
func()	Function handler to invoke that is specific to the ioctl.
name	Debug character string


Ioctl format:
All ioctls start with the same basic header, followed by the same
data layout.

uda_ioctl
=========
ns_id	Identifies the name space which is the target of this ioctl.
	The name space manager is always at id 0.
length	Total size of the ioctl.  Note that additional input/output
	buffers may be provided as arguments, which would not be
	included as part of this length.
op	The actual operation to perform.  For debugging purposes,
	op values are not shared between name spaces.  The op
	is used as a simple index into a name space's descriptor table.
flags	Any flags associated with the operation.
obj_cnt	Kernel allocated objects follow the ioctl header.  This
	specifies the number of objects.
arg_cnt	A number of input/output arguments following any objects.
data	Ioctl data, including objects and arguments follow the
	header.  Notable data include the uda_obj_id and uda_arg
	structures


uda_obj_id
==========
instance_id
	An index into an idr, which represents a kernel object.
obj_type
	Identifies the type of object represented by the instance_id.
	The framework validates the obj_type against that stored by
	the kernel.
data	Object specific information (padding)

uda_arg
=======
offset	Byte offset from the start of the ioctl where the argument
	data begins.
attr_id	A unique identifier for the ioctl data.  The attribute id
	typically corresponds to a data structure.  Similar to
	opcodes, the attribute id's are not shared between name
	spaces.
length	Size of the argument data
data	Attribute specific information (padding)


This is the general flow of ioctl processing through the framework.

1. Copy ioctl header from user space.
2. Verify length of ioctl, including arg sizes.
3. Select target name space.
4. Name space verifies ioctl and returns descriptor.
5. Copy ioctl argument data from user space.
6. Verify and convert user space object identifiers to kernel
   objects.
7. Invoke ioctl handler
8. Return control to user space.

There are some kinks to work out of this, but I think it provides a
generic enough framework to support current and anticipated needs,
and seems to align with the work being presented in the ofvwg.  A
generic attribute validation scheme could easily fit into this
under the get_ioctl_desc() callback.  (I almost have the start of
one if you examine the patch details.)  Alternatively, by
defining a 'compat' attribute id, the existing write interface could
probably be ported almost as-is, though I'm not recommending this.

The event interface needs to be defined.  But if each device is
treated as a new name space, I think this will allow a single fd to
issue commands to any rdma device.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>